Correct errors in preperation for merging
Just a few issues. First, I'm not entirely confident in changing the Omnis gateway to the red wisp, rather, I have added solar and fuel collection to the sprite for Omnis ringworlds, a value that would equal roughly 1 in total.

I have also fixed an issue with the peregrine sprite definition and offset, so that it actually shows up in the arena.