Update Carrier and Aerie, and fix old Wardragon (#85)
The old Quarg Wardragon tribute fleet spawned the new Quarg Wardragon. This was fixed so that it spawns the old Quarg Wardragon, as expected.

Furthermore, both the Aerie and Carrier tributes do not spawn the full complement of fighters, as they were not updated when the Aerie and Carrier were given more bays. Updated the tribute fleets of the Aerie and the Carrier to fix this.